+++ source/navier_stokes_projection/assemble_projection_rhs.cc
@@ -17,7 +17,7 @@ assemble_projection_step_rhs()
 
   const int p_degree = velocity.fe_degree + pressure.fe_degree - 1;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
